requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
sqlalchemy==2.0.43
asyncpg==0.30.0
alembic==1.16.4
aiogram==3.22.0
httpx==0.28.1
python-dotenv==1.1.1